pyproject.toml
[project]
name = "assistant-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "greenlet==3.2.3",
    "langchain==0.3.25",
    "langchain-community==0.3.25",
    "langchain-mcp-adapters>=0.1.8",
    "langchain-openai>=0.3.27",
    "langchain-tavily>=0.2.0",
    "langgraph==0.4.8",
    "langgraph-prebuilt==0.2.2",
    "langmem>=0.0.28",
    "python-dotenv>=1.1.1",
    "aiogram==3.21.0"
]
